Remove broken notification on auth prompt
When the window pops up there is no way to tell immediately if it is
actually the top level window or not, which means we always send a
notification that is instantly dismissed, resulting in a weird flicker.

Even if we could tell there's no way to be sure the WM hasn't switched
it before the user actually sees it.

Removed for now as in future this prompt will move into kwin/plasmashell
anyway where we can control these things, so there's no point spending
too much time on it.