Use .cmake for dynamically generated desktop file
The translation infrastructure extracts messages from dynamically-generated
desktop files (from CMake) when they are called desktop.cmake.

It would be possible to translate the content of this file even if it was
called .desktop.in, but it requires more configuration, so let's follow
the established process.

Those messages have been available but it was not possible to translate them
for some years, so adding them back even in the stable branch is not a break
of the string freeze.