Always use "lock" icon in auth dialog, disregarding icon from auth action
In the past, the old QtWidgets-based auth dialog disregarded the icon
set by auth actions in favor of a hardcoded one. This was presumably
because we wanted to always show an appropriate and consistent icon,
instead of whatever random and inconsistent icon was specified by all
the different auth actions out there.

As a result, we got sloppy about setting icons in our own auth action
files. Many are using the generic/default `kauth-example` icon name
which does not exist in the Breeze theme.

However the new Plasma 6 based dialog does honor the icon in the auth
file. As a result, the dialog shows a lot of broken and inconsistent
icons.

We could fix this by changing the icons in all of our auth files
everywhere. However that has two downsides:
1. It would be a lot of work
2. The result would still most likely result in icon inconsistency, when
   all we really want to do is show the same lock-style icon in the
   dialog every time

More work for a worse result doesn't strike me as an ideal use of time.
Thus, this commit goes in the other direction and brings back the
approach taken by the old QtWidgets dialog by ignoring the icon set by
auth actions, instead always showing the "lock" icon in the dialog.

BUG: 473720