Extend service start timeout
On a slow machine the service may take longer than 5 seconds (walltime) to start,
especially during session start when multiple processes are competing for
CPU time and I/O.

The original reasoning for the timeout is to limit the time to quit the process
when the session is shut down, but that can be set with TimeoutStopSec.

Remove the start timeout, as it os questionable if is useful at all. It only
helps if there is some deadlock during startup which would be cured by a
fresh start, otherwise it is just consuming CPU time again and again.

For comparison, Gnome does not use Timeout[Start]Sec at all, but
only TimeoutStopSec.

BUG: 477805