Fix crash on hyprland
In policykitlistener.cpp, we check to see if m_session is null.
Regardless of whether it was or wasn't, we later call
m_session.data()->deleteLater(), which explodes if it was.

Only do this if it was not null.

BUG: 485407
FIXED-IN: 6.1


(cherry picked from commit de434550e7c41ac685e42b35b9d076a6948e7bc2)